# Optimize metering flushes