@graydon first experimented and identified this issue. I'm making the code change and presenting findings. 

**What** 
Step meter's `flush` is implicitly called by the meter itself when the instruction count reaches the threshold (256). 
This PR ensures: only explicitly call `flush` if we are returning from WASM to host, either 1. the entire code execution finishes and returns 2. branches to function call which may return to host. 

**Why**
Previously, we were flushing in every `br` call that are still in the same wasm execution context. This results in premature meter charging (where the threshold is not respected), and slower execution. 

I ran the [fannkuch](https://github.com/stellar/rs-soroban-env/tree/main/soroban-test-wasms/wasm-workspace/fannkuch) contract (around 38M steps, where each step is an interpreter loop) and did a comparison and observed significant improvements (~33%) in performance.
|| Before      | After |
| -----------|----------- | ----------- |
| No. flushes |   4.31M    | 140k   |
| Ave. steps per flush | 8   | 256        |
|Exec. time| 128ms | 85ms |
|Exec.time (nsecs)/step| 3 | 2 |