@graydon 2022-11-23
@jayz22 nice! could you check at a few other flush-granularities, just to get an idea of where we start to see diminishing returns?