@jayz22 2022-11-23
> @jayz22 nice! could you check at a few other flush-granularities, just to get an idea of where we start to see diminishing returns?

I'm going to experiment with the step size next. 
Given the current step size of 256, the current flush granularity is optimal. Any additional flush call (i.e. at `Br`, `BrIfEqz`, `BrIfNez`) is premature which results in reduced effective step size, and slows down the execution. See results below:

|| Optimal (this PR)      | this + `Br` | this + `Br` + `BrIfEqz` | this + `Br`+ `BrIfEqz` + `BrIfNez` |
| ------------------------|------------------------ | ------------------------ | ------------------------|------------------------|
| No. flushes |   149978    | 942086   | 2875577 |  4318978 | 
| Ave. steps per flush | 255   | 40  | 13 | 8 | 
|Exec. time (ms)| 95 | 100 | 125 | 140 |
|Exec.time (nsecs)/step| 2.48 |  2.60 | 3.26 | 3.66 |

Does this answer your question @graydon ?