@graydon 2022-11-23
@jayz22 oh, I meant varying the step size! But this is also interesting