@jayz22 2022-11-23
I have the results for the step size experiment in the attached PDF. 
[result_sorted.pdf](https://github.com/stellar/wasmi/files/10079411/result_sorted.pdf)

Key results:
- The performance (cpu insns and time) follows a nice inverse proportion relation to the input size, as expected.
- The last column is the cpu insns normalized (between 0 and 1). A number of 0.1 means 10% aways from the baseline number (or 90% "good"). I've highlighted a few key numbers in the table below, corresponding to 20%, 10%, 5%, 2.3% (our current step size of 256), and 1%. 

step size | cpu insns | time nsecs | insns/input | nsecs/input | norm cpu
-- | -- | -- | -- | -- | --
40 | 1106590710 | 101991250 | 28.8 | 2.65 | 0.19364141487135000
72 | 1021274141 | 96017750 | 26.5 | 2.5 | 0.1040849955668470
136 | 971160225 | 92680833 | 25.2 | 2.41 | 0.05148064301386730
256 | 944625240 | 90308916 | 24.6 | 2.35 | 0.02362698843360990
448 | 931760456 | 89124333 | 24.2 | 2.32 | 0.010122882447281900

I think depending on the goal, if we want a more conservative approach: choosing a finer step so that we don't miss any metering, then I think 72 would be good (90%).
If we want better performance, then I think 448 (99%) would be a good choice. 
@graydon  let me know what you think.